feat: 유저제작 문제, 세트 코드작성

+++ codeBoy_backend/src/main/java/com/codeboy/mvc/model/dao/ScoreDao.java
@@ -12,7 +12,7 @@ public interface ScoreDao {
 	public List<UserScore> selectAllUserScore();
 
 	//한 멤버의 스코어 조회
-	public int selectOneUserScore(int memberId);
+	public int selectOneUserScore(Long memberId);
 
 	//멤버 스코어 업데이트
 	public void updateUserScore(UserScore userScore);
